Removing more JVM references.
