+++ core/currency/build.gradle.kts
@@ -13,6 +13,12 @@ kotlin {
                 implementation(kotlin("test"))
             }
         }
+        val androidMain by getting
+        val androidUnitTest by getting {
+            dependencies {
+                dependsOn(commonTest)
+            }
+        }
         maybeCreate("iosX64Main")
         maybeCreate("iosArm64Main")
         maybeCreate("iosSimulatorArm64Main")
