+++ core/currency/src/androidMain/kotlin/com/adammcneilly/pocketleague/core/currency/AndroidCurrencyFormatter.kt
@@ -1,0 +1,17 @@
+package com.adammcneilly.pocketleague.core.currency
+
+import java.text.NumberFormat
+import java.util.Currency
+
+/**
+ * A concrete implementation of [CurrencyFormatter] to use on Android platform.
+ */
+class AndroidCurrencyFormatter : CurrencyFormatter {
+
+    override fun formatCurrency(amount: Double, currency: String): String? {
+        val formatter = NumberFormat.getCurrencyInstance()
+        formatter.currency = Currency.getInstance(currency)
+
+        return formatter.format(amount)
+    }
+}
