+++ core/currency/src/androidMain/kotlin/com/adammcneilly/pocketleague/core/currency/CurrencyFormatter.android.kt
@@ -1,0 +1,8 @@
+package com.adammcneilly.pocketleague.core.currency
+
+/**
+ * Returns the [CurrencyFormatter] instance for this target.
+ */
+actual fun currencyFormatter(): CurrencyFormatter {
+    return AndroidCurrencyFormatter()
+}
