+++ core/locale/src/androidTest/kotlin/com/adammcneilly/pocketleague/core/locale/LocaleHelperTest.jvm.kt
@@ -4,7 +4,7 @@ import com.varabyte.truthish.assertThat
 import kotlin.test.Test
 
 class LocaleHelperTest {
-    private val localeHelper = JVMLocaleHelper()
+    private val localeHelper = AndroidLocaleHelper()
 
     @Test
     fun getCountryDisplayName() {
