+++ data/remote/src/androidMain/kotlin/com/adammcneilly/pocketleague/data/remote/HTTPClientEngine.android.kt
@@ -1,0 +1,11 @@
+package com.adammcneilly.pocketleague.data.remote
+
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.cio.CIO
+
+/**
+ * Return a [CIO] [HttpClientEngine] for the Android platform.
+ */
+actual fun provideHttpClientEngine(): HttpClientEngine {
+    return CIO.create()
+}
